checked before the tour is deleted

+++ src/main/java/ua/tour/api/services/TourService.java
@@ -4,26 +4,34 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ua.tour.api.entities.Tour;
 import ua.tour.api.repo.TourRepository;
+import ua.tour.api.repo.TourReservationRepository;
 
-import java.util.Date;
 import java.util.Optional;
 
 @Service
 public class TourService {
 
     private final TourRepository tourRepository;
+    private final TourReservationRepository reservationRepository;
 
     @Autowired
-    public TourService(TourRepository tourRepository) {
+    public TourService(TourRepository tourRepository, TourReservationRepository tourReservationRepository) {
         this.tourRepository = tourRepository;
+        this.reservationRepository = tourReservationRepository;
     }
 
     public void createNewTour(Tour tour) {
         tourRepository.save(tour);
     }
 
-    public void removeTour(Long tourId) {
-        tourRepository.deleteById(tourId);
+    public void removeTour(Long tourId) throws Exception {
+        Optional<Tour> opTour = tourRepository.findById(tourId);
+        if (opTour.isPresent()) {
+            Long reservationsCount = reservationRepository.countByTour(opTour.get());
+            if (reservationsCount == 0) {
+                tourRepository.deleteById(tourId);
+            } else throw new Exception();
+        } else throw new Exception();
     }
 
 
@@ -31,9 +39,9 @@ public class TourService {
         return tourRepository.findAll();
     }
 
-    public void updateTour(Tour tour) {
+    public void updateTour(Tour tour) throws Exception {
         if (tourRepository.existsById(tour.getId())) {
             tourRepository.save(tour);
-        }
+        } else throw new Exception();
     }
 }
